feat(language-server): Use the default path defined in contextive if it's not specified in the configuration.
This enhancement to the language server allows an empty path
configuration to fall back to a default location:
`.contextive/definitions.yml`.

Introduce a new literal value called `defaultContextiveDefinitionsPath`
to serve as a fallback for the "path" configuration when it's not set.

Modify the test "Server fails to load contextive file with an error when
no configuration is supplied" in `InitializationTests` to align with the
new functionality.