feat(language-server): use a config variable to store configuration changes
When using for example the neovim language server client, the contextive
language server does not acquire the proper settings sent by the client
in the `OnStarted` handler. However, the settings are passed correctly
using the `OnDidChangeConfiguration` handler.

This change introduces a scoped variable in the server start function to
hold a mutable field for the Contextive path configuration. The
configuration variable is initialized with the default path for
Contextive definitions introduced in commit 23f9049.

If the language configuration does not have a config value, it will use
the value from the configuration variable instead. This works the same
way as it did previously, but the source is now a parameter of the
function.