test(vscode): fix creation of spurious folders when compiling vscode tests
When fable is run on a folder with the cwd in a parent folder, for some reason when it finds .cs files it looks for those files relative to the cwd.  When not found, the folders they are in get created, leading to strange spurious folders being created - specifically, the single-root folder in the main vscode/contextive folder, due to the .cs files in the test fixtures.